+++ messages/KRBPriv.go
@@ -22,7 +22,7 @@ type EncKrbPrivPart struct {
 	UserData       []byte            `asn1:"explicit,tag:0"`
 	Timestamp      time.Time         `asn1:"generalized,optional,explicit,tag:1"`
 	Usec           int               `asn1:"optional,explicit,tag:2"`
-	SequenceNumber int               `asn1:"optional,explicit,tag:3"`
+	SequenceNumber int64             `asn1:"optional,explicit,tag:3"`
 	SAddress       types.HostAddress `asn1:"explicit,tag:4"`
 	RAddress       types.HostAddress `asn1:"optional,explicit,tag:5"`
 }
